requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-telegram-bot==21.5
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2